pyproject.toml
[tool.poetry]
name = "attack-sync"
version = "0.1.0"
description = ""
authors = ["Center for Threat-Informed Defense <ctid@mitre-engenuity.org>"]
license = "Apache"
readme = "README.md"

[tool.poetry.scripts]
build-attack-sync = "attack_sync.site_builder:main"

[tool.poetry.dependencies]
python = "^3.10"
openpyxl = "^3.1.2"
setuptools = "^75.1.0"
jinja-markdown2 = "^0.4.0"
markupsafe = "^3.0.2"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
markdown = "^3.4.3"
loguru = "^0.7.0"
requests = "^2.28.2"
deepdiff = "^6.3.0"
python-dateutil = "^2.8.2"
jinja2 = "^3.1.2"
rich = "^13.3.4"
stix2 = "^3.0.1"
tqdm = "^4.65.0"
mitreattack-python = "^3.0.8"
mypy = "^1.2.0"
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
skip_gitignore = true
